requirements.txt
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-limiter==0.1.6
youtube-transcript-api==1.1.1
google-generativeai==0.8.5
httpx==0.27.0
redis==6.2.0
python-dotenv==1.1.0
uvicorn==0.34.3